feat(video-overlay): split VideoOverlay component into modular parts
- Refactored VideoOverlay into three new components:
  - VideoControlButton: Handles individual video control button styling and logic
  - VideoControls: Contains all video control buttons, handling their layout and state
  - VideoPlayer: Manages the video player display and playback functionality
- Updated component structure to improve readability and maintainability.
- Adjusted styling with Tailwind CSS for better modularity and reuse.

Refactor allows for easier testing and flexibility in future feature extensions.
