--- src/components/overlays/VideoOverlay.tsx
@@ -1,209 +1,0 @@
-import React, { useState, useRef, useEffect, useCallback } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faPlay, faPause, faRedo } from '@fortawesome/free-solid-svg-icons';
-
-interface VideoOverlayProps {
-  videoLink: string;
-  onClose: () => void;
-}
-
-const VideoOverlay: React.FC<VideoOverlayProps> = ({ videoLink, onClose }) => {
-  const baseUrl = "assets/videos";
-  const isYouTubeLink = (url: string) => {
-    return url.includes('youtube.com') || url.includes('youtu.be');
-  };
-
-  const getYouTubeEmbedUrl = (url: string) => {
-    const videoIdMatch = url.match(/(?:youtube\.com\/(?:[^/]+\/.+\/|(?:v|e(?:mbed)?)\/|.*[?&]v=)|youtu\.be\/)([^"&?/\\s]{11})/);
-    return videoIdMatch ? `https://www.youtube.com/embed/${videoIdMatch[1]}` : '';
-  };
-
-  const videoRef = useRef<HTMLVideoElement | null>(null);
-  const containerRef = useRef<HTMLDivElement | null>(null);
-  const [isPlaying, setIsPlaying] = useState(false);
-  const [progress, setProgress] = useState(0);
-  const [longPressTimeout, setLongPressTimeout] = useState<NodeJS.Timeout | null>(null);
-
-  // Handle progress update
-  const updateProgress = () => {
-    if (videoRef.current) {
-      const currentTime = videoRef.current.currentTime;
-      const duration = videoRef.current.duration;
-      setProgress((currentTime / duration) * 100);
-    }
-  };
-
-  // Toggle play/pause
-  const togglePlayPause = () => {
-    if (videoRef.current) {
-      if (isPlaying) {
-        videoRef.current.pause();
-      } else {
-        videoRef.current.play();
-      }
-      setIsPlaying(!isPlaying);
-    }
-  };
-
-  // Reset video
-  const resetVideo = () => {
-    if (videoRef.current) {
-      videoRef.current.pause(); // Pause the video if it's playing
-      videoRef.current.currentTime = 0;
-      setProgress(0);
-      setIsPlaying(false);
-    }
-  };
-
-  // Seek video
-  const seekVideo = (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    if (videoRef.current) {
-      const rect = event.currentTarget.getBoundingClientRect();
-      const offsetX = event.clientX - rect.left;
-      const newTime = (offsetX / rect.width) * videoRef.current.duration;
-      videoRef.current.currentTime = newTime;
-    }
-  };
-
-  // Close overlay if clicked outside the video container
-  const handleClickOutside = useCallback((event: MouseEvent) => {
-    if (containerRef.current && !containerRef.current.contains(event.target as Node)) {
-      onClose();
-    }
-  }, [onClose]);
-
-  useEffect(() => {
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [handleClickOutside]);
-
-  // Handle keyboard events for play/pause and exit
-  const handleKeyDown = useCallback((event: KeyboardEvent) => {
-    if (event.key === 'Escape') {
-      onClose(); // Exit the video overlay
-    } else if (event.key === ' ') {
-      event.preventDefault(); // Prevent scrolling
-      togglePlayPause(); // Toggle play/pause
-    } else if (event.key.toLowerCase() === 'r') {
-      resetVideo(); // Reset video
-    }
-  }, [onClose, togglePlayPause]);
-
-  useEffect(() => {
-    document.addEventListener('keydown', handleKeyDown);
-    return () => {
-      document.removeEventListener('keydown', handleKeyDown);
-    };
-  }, [handleKeyDown]);
-
-  // Handle long press to close the overlay
-  const handleLongPressStart = () => {
-    const timeout = setTimeout(() => {
-      onClose();
-    }, 500); // 500ms for long press
-    setLongPressTimeout(timeout);
-  };
-
-  const handleLongPressEnd = () => {
-    if (longPressTimeout) {
-      clearTimeout(longPressTimeout);
-      setLongPressTimeout(null);
-    }
-  };
-
-  return (
-    <div className={`fixed inset-0 flex items-center justify-center bg-black bg-opacity-75 z-20 transition-opacity duration-300 ease-in-out`}>
-      <div
-        ref={containerRef}
-        className={`relative w-full h-full max-w-5xl max-h-full flex items-center justify-center p-4 bg-black`}
-        onTouchStart={handleLongPressStart}
-        onTouchEnd={handleLongPressEnd}
-        onMouseDown={handleLongPressStart} // For desktop long press
-        onMouseUp={handleLongPressEnd} // For desktop long press
-      >
-        <button
-          onClick={onClose}
-          className="absolute top-4 right-4 bg-red-600 text-white rounded-full w-14 h-14 flex items-center justify-center z-40"
-          aria-label="Close video"
-        >
-          X
-        </button>
-
-        <div className="w-full h-auto aspect-w-16 aspect-h-9 max-w-full max-h-full bg-gray-800 rounded-lg overflow-hidden relative">
-          {isYouTubeLink(videoLink) ? (
-            <iframe
-              className="w-full h-full"
-              src={getYouTubeEmbedUrl(videoLink)}
-              title="YouTube video player"
-              allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-              allowFullScreen
-              style={{ border: 0 }}
-            ></iframe>
-          ) : (
-            <video
-              ref={videoRef}
-              controls={false}
-              className="object-contain w-full h-full"
-              onClick={togglePlayPause}
-              onTimeUpdate={updateProgress}
-            >
-              <source src={`${baseUrl}/${encodeURIComponent(videoLink)}`} type="video/mp4" />
-              <source src={`${baseUrl}/${encodeURIComponent(videoLink.replace('.mp4', '.webm'))}`} type="video/webm" />
-              <source src={`${baseUrl}/${encodeURIComponent(videoLink.replace('.mp4', '.ogv'))}`} type="video/ogg" />
-              Your browser does not support the video tag.
-            </video>
-          )}
-          <div className="absolute bottom-0 left-0 w-full h-2 bg-gray-600 cursor-pointer" onClick={seekVideo}>
-            <div className="h-full bg-red-500" style={{ width: `${progress}%` }}></div>
-          </div>
-        </div>
-
-        {/* Control Buttons Inside the Video Box */}
-        <div className="absolute z-50 bottom-4 left-1/2 transform -translate-x-1/2 flex space-x-4">
-          {/* Play/Pause Button */}
-          <button
-            onClick={togglePlayPause}
-            className={`hidden lg:flex items-center justify-center bg-${isPlaying ? 'red-500' : 'green-500'} text-white rounded-lg px-4 py-2 shadow-lg hover:bg-opacity-80 transform transition-transform duration-150 ease-in-out`}
-            aria-label={isPlaying ? 'Pause video' : 'Play video'}
-          >
-            <FontAwesomeIcon icon={isPlaying ? faPause : faPlay} className="mr-2" />
-            {isPlaying ? 'Pause' : 'Play'}
-          </button>
-          {/* Reset Button */}
-          <button
-            onClick={resetVideo}
-            className="hidden lg:flex items-center justify-center bg-blue-500 text-white rounded-lg px-4 py-2 shadow-lg hover:bg-opacity-80 transform transition-transform duration-150 ease-in-out"
-            aria-label="Reset video"
-          >
-            <FontAwesomeIcon icon={faRedo} className="mr-2" />
-            Reset
-          </button>
-        </div>
-
-        {/* Mobile Control Buttons */}
-        <div className="absolute top-4 left-4 flex flex-col space-y-2 z-50 md:hidden">
-          {/* Play/Pause Button */}
-          <button
-            onClick={togglePlayPause}
-            className={`flex items-center justify-center bg-${isPlaying ? 'red-500' : 'green-500'} text-white rounded-lg p-2 shadow-lg hover:bg-opacity-80 transform transition-transform duration-150 ease-in-out`}
-            aria-label={isPlaying ? 'Pause video' : 'Play video'}
-          >
-            <FontAwesomeIcon icon={isPlaying ? faPause : faPlay} />
-          </button>
-          {/* Reset Button */}
-          <button
-            onClick={resetVideo}
-            className="flex items-center justify-center bg-blue-500 text-white rounded-lg p-2 shadow-lg hover:bg-opacity-80 transform transition-transform duration-150 ease-in-out"
-            aria-label="Reset video"
-          >
-            <FontAwesomeIcon icon={faRedo} />
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default VideoOverlay;
